+++ ui.R
@@ -51,6 +51,8 @@ shinyUI(bootstrapPage(
                         tabPanel("Import networks",                        
                             checkboxInput('chkEdgeListFileHeader', 'Header', FALSE),
                             textInput("txtEdgeListFileSep", label=HTML("Separator (default is one space):"), " "),
+                            helpText(HTML("<font color='red'>If nodes in edges lists are identified by their label instead of a sequential integer ID, check the box below</font> (note that you must specify sequential integer IDs in the nodeID column of the layout file)")),
+                            checkboxInput("chkEdgeListLabel", "Edges list of labeled nodes (instead of sequential integer)", FALSE),
                             selectInput("selEdgeListType", HTML("<strong>* EdgeList Format:</strong>"), 
                                 choices = c("Undirected", "Directed")),
                             checkboxInput("chkEdgeListWeighted", "Weighted", FALSE),